Add model subresources - (Task #952)
For sei, ca and property. Also add helper class.

---
Task #952: Refactor ModelAccessResource
